fix(code-generation/go): support a typeSwitch nested in a typeSwitch case
Two gaps kept the Go output from compiling as soon as a mspec nests a
typeSwitch inside a case and dispatches on a discriminator the intermediate
type declares itself. eip.mspec is the first live user of that construct,
which is why this went unnoticed.

1) The intermediate's interface lacked the discriminator accessor its own
   serializer calls, producing "m.GetService undefined". The template asked
   ComplexTypeDefinition.isDiscriminatorField(), which resolves the names
   from getParentType().orElse(this) and therefore only ever sees the
   parent's discriminators - while getDiscriminatorTypes() already prefers
   the type's own switch field, so the two disagreed. Rather than change
   that shared semantic (it would affect every language template, and Java
   is correct today), ask the type directly through the new
   declaresDiscriminatorField helper.

2) Child parse calls passed their arguments bare. Go has no implicit numeric
   conversion, so an argument whose declared width differs from the value
   bound to it (a uint8 field bound to a uint16 parser argument) does not
   compile. Integer arguments are now converted to the declared type, which
   is a no-op wherever the widths already agree - hence the regenerated
   models only gain uint8(x)/uint16(x) wrappers.

+++ plc4go/protocols/abeth/readwrite/model/CIPEncapsulationPacket.go
@@ -446,7 +446,7 @@ func (m *_CIPEncapsulationPacket) parse(ctx context.Context, readBuffer utils.Re
 			return nil, errors.Wrap(err, "Error parsing sub-type CIPEncapsulationReadRequest for type-switch of CIPEncapsulationPacket")
 		}
 	case commandType == 0x0207: // CIPEncapsulationReadResponse
-		if _child, err = new(_CIPEncapsulationReadResponse).parse(ctx, readBuffer, m, packetLen); err != nil {
+		if _child, err = new(_CIPEncapsulationReadResponse).parse(ctx, readBuffer, m, uint16(packetLen)); err != nil {
 			return nil, errors.Wrap(err, "Error parsing sub-type CIPEncapsulationReadResponse for type-switch of CIPEncapsulationPacket")
 		}
 	default:
